# improve parsing & serialization of collections