Results of benchmark for collections.
Before:
```
[info] SerializerBenchmark.writeThanReadIterables                                    thrpt    5   967112.118 ±  451579.431   ops/s
[info] SerializerBenchmark.writeThanReadIterables:·gc.alloc.rate                     thrpt    5      609.798 ±     284.751  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.alloc.rate.norm                thrpt    5      992.000 ±       0.001    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Eden_Space            thrpt    5      601.759 ±     929.951  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Eden_Space.norm       thrpt    5      968.227 ±    1279.809    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Survivor_Space        thrpt    5        0.021 ±       0.113  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Survivor_Space.norm   thrpt    5        0.032 ±       0.174    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.count                          thrpt    5        7.000                counts
[info] SerializerBenchmark.writeThanReadIterables:·gc.time                           thrpt    5       10.000                    ms
[info] SerializerBenchmark.writeThanReadMaps                                         thrpt    5  1319055.847 ±   24233.237   ops/s
[info] SerializerBenchmark.writeThanReadMaps:·gc.alloc.rate                          thrpt    5      831.745 ±      15.290  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.alloc.rate.norm                     thrpt    5      992.000 ±       0.001    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Eden_Space                 thrpt    5      824.315 ±     387.600  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Eden_Space.norm            thrpt    5      983.292 ±     467.485    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Survivor_Space             thrpt    5        0.083 ±       0.113  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Survivor_Space.norm        thrpt    5        0.099 ±       0.137    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.count                               thrpt    5       11.000                counts
[info] SerializerBenchmark.writeThanReadMaps:·gc.time                                thrpt    5       23.000                    ms
```
After these changes applied (on the same env):
```
[info] SerializerBenchmark.writeThanReadIterables                                    thrpt    5  1895629.080 ±  51028.891   ops/s
[info] SerializerBenchmark.writeThanReadIterables:·gc.alloc.rate                     thrpt    5      549.431 ±     14.770  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.alloc.rate.norm                thrpt    5      456.000 ±      0.001    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Eden_Space            thrpt    5      598.037 ±    867.386  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Eden_Space.norm       thrpt    5      496.031 ±    715.509    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Survivor_Space        thrpt    5        0.029 ±      0.108  MB/sec
[info] SerializerBenchmark.writeThanReadIterables:·gc.churn.PS_Survivor_Space.norm   thrpt    5        0.024 ±      0.089    B/op
[info] SerializerBenchmark.writeThanReadIterables:·gc.count                          thrpt    5        7.000               counts
[info] SerializerBenchmark.writeThanReadIterables:·gc.time                           thrpt    5        7.000                   ms
[info] SerializerBenchmark.writeThanReadMaps                                         thrpt    5  2656908.927 ±  75533.348   ops/s
[info] SerializerBenchmark.writeThanReadMaps:·gc.alloc.rate                          thrpt    5      607.961 ±     17.352  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.alloc.rate.norm                     thrpt    5      360.000 ±      0.001    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Eden_Space                 thrpt    5      614.763 ±    279.855  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Eden_Space.norm            thrpt    5      364.021 ±    164.494    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Survivor_Space             thrpt    5        0.075 ±      0.108  MB/sec
[info] SerializerBenchmark.writeThanReadMaps:·gc.churn.PS_Survivor_Space.norm        thrpt    5        0.044 ±      0.064    B/op
[info] SerializerBenchmark.writeThanReadMaps:·gc.count                               thrpt    5       14.000               counts
[info] SerializerBenchmark.writeThanReadMaps:·gc.time                                thrpt    5       38.000                   ms
```