@nau 2017-09-28
Would be great to see a comparison with reflection-based Kryo serialization.