@plokhotnyuk 2017-09-28
@nau do you mean chill-scala https://github.com/twitter/chill or something else?