@nau 2017-09-29
Chill would be great, but I meant https://github.com/EsotericSoftware/kryo#default-serializers, Kryo's default reflection-based FieldSerializer.
The idea is to show how much macros serializers faster than reflection-based serializers with almost none additional boilerplate required.